+++ pkg/binder/common/constants.go
@@ -5,7 +5,6 @@ package common
 
 const (
 	NvidiaVisibleDevices = "NVIDIA_VISIBLE_DEVICES"
-	VisibleDevices       = "RUNAI-VISIBLE-DEVICES"
 	NumOfGpusEnvVar      = "RUNAI_NUM_OF_GPUS"
 	ReceivedTypeFraction = "Fraction"
 	ReceivedTypeRegular  = "Regular"