+++ CHANGELOG.md
@@ -6,14 +6,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [v0.7.7] - 2025-07-16
+
+### Fixed
+- Fixed a scenario where only GPU resources where checked for job and node, causing it to be bound instead of being pipelined
+
+## [v0.7.6] - 2025-07-11
+
 ### Added
 - Added GPU_PORTION env var for GPU sharing pods
 
-### Changed
-- Changed RUNAI-VISIBLE-DEVICES key in GPU sharing configmap to NVIDIA_VISIBLE_DEVICES
+## [v0.7.5] - 2025-07-10
 
 ### Fixed
 - Fixed a miscalculation where cpu/memory releasing resources were considered idle when requesting GPU fraction/memory
+
+## [v0.7.4] - 2025-07-09
+
+### Changed
+- Changed RUNAI-VISIBLE-DEVICES key in GPU sharing configmap to NVIDIA_VISIBLE_DEVICES
 
 ## [v0.7.3] - 2025-07-08